chore: use Extend.Admin
